Rename `lib` -> `core`
