+++ android/src/main/java/com/swmansion/gesturehandler/core/FlingGestureHandler.kt
@@ -1,4 +1,4 @@
-package com.swmansion.gesturehandler.lib
+package com.swmansion.gesturehandler.core
 
 import android.os.Handler
 import android.os.Looper
