+++ android/src/main/java/com/swmansion/gesturehandler/core/GestureHandlerInteractionController.kt
@@ -1,4 +1,4 @@
-package com.swmansion.gesturehandler.lib
+package com.swmansion.gesturehandler.core
 
 interface GestureHandlerInteractionController {
   fun shouldWaitForHandlerFailure(handler: GestureHandler<*>, otherHandler: GestureHandler<*>): Boolean
