+++ android/src/main/java/com/swmansion/gesturehandler/core/GestureHandlerRegistry.kt
@@ -1,4 +1,4 @@
-package com.swmansion.gesturehandler.lib
+package com.swmansion.gesturehandler.core
 
 import android.view.View
 import java.util.*
