+++ android/src/main/java/com/swmansion/gesturehandler/core/PointerEventsConfig.kt
@@ -1,4 +1,4 @@
-package com.swmansion.gesturehandler.lib
+package com.swmansion.gesturehandler.core
 
 enum class PointerEventsConfig {
   /**
