+++ android/src/main/java/com/swmansion/gesturehandler/core/RotationGestureHandler.kt
@@ -1,8 +1,8 @@
-package com.swmansion.gesturehandler.lib
+package com.swmansion.gesturehandler.core
 
 import android.graphics.PointF
 import android.view.MotionEvent
-import com.swmansion.gesturehandler.lib.RotationGestureDetector.OnRotationGestureListener
+import com.swmansion.gesturehandler.core.RotationGestureDetector.OnRotationGestureListener
 import kotlin.math.abs
 
 class RotationGestureHandler : GestureHandler<RotationGestureHandler>() {
