+++ android/src/main/java/com/swmansion/gesturehandler/react/RNGestureHandlerEvent.kt
@@ -11,7 +11,7 @@ import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.WritableMap
 import com.facebook.react.uimanager.events.Event
 import com.facebook.react.uimanager.events.RCTEventEmitter
-import com.swmansion.gesturehandler.lib.GestureHandler
+import com.swmansion.gesturehandler.core.GestureHandler
 
 class RNGestureHandlerEvent private constructor() : Event<RNGestureHandlerEvent>() {
   private var extraData: WritableMap? = null
