+++ android/src/main/java/com/swmansion/gesturehandler/react/RNGestureHandlerEventDataExtractor.kt
@@ -1,7 +1,7 @@
 package com.swmansion.gesturehandler.react
 
 import com.facebook.react.bridge.WritableMap
-import com.swmansion.gesturehandler.lib.GestureHandler
+import com.swmansion.gesturehandler.core.GestureHandler
 
 interface RNGestureHandlerEventDataExtractor<T : GestureHandler<T>> {
   fun extractEventData(handler: T, eventData: WritableMap)
